clinker-core: extract Aggregation dispatch arm to executor/aggregate_dispatch.rs
Move the PlanNode::Aggregation arm body — hash / streaming strategy
selection, per-record ingest, spill triggers, finalize, and the
relaxed-CK retraction handoff — out of executor/dispatch.rs into a new
sibling module executor/aggregate_dispatch.rs, together with the
time-windowed aggregation helpers it owns: run_time_windowed_aggregate,
add_to_window, handle_aggregate_add_error, finalize_windows,
push_late_record, emit_aggregate_finalize_dlq. Those helpers were already
scoped entirely to the Aggregation arm. The dispatcher's Aggregation arm
is now a single delegating call into aggregate_dispatch::dispatch_aggregation.

Pure relocation: no symbol renames, no public-API change. The moved body
calls cross-arm helpers (record_error_to_buffer_if_grouped,
tee_emit_to_region_input_buffers, stream_linear_producer_emit,
project_rows_to_buffer_schema) and shared types (RetainedAggregatorState)
that stay in dispatch.rs and remain pub(crate).

Closes #246
